fix(ci): shard6 verdict precedence and shard8 line-limit — preserve reasoned NEEDS_HUMAN and split oversized test module
